Added tests for sesson filters
