+++ bn-apps/cordapp-updates-distribution/corda-updates-app/src/test/resources/corda-updates-app-with-filter.conf
@@ -1,0 +1,5 @@
+configPath = "./config.conf"
+syncInterval = 9999999
+notary = "O=Notary,L=London,C=GB"
+bno = "O=BNO,L=London,C=GB"
+sessionFilter = "net.corda.cordaupdates.app.DenyAllSessionFilter"
