+++ bn-apps/cordapp-updates-distribution/corda-updates-app/src/test/resources/corda-updates-app.conf
@@ -1,4 +1,4 @@
-configPath=./config.conf
-syncInterval=9999999
-notary="O=Notary,L=London,C=GB"
-bno="O=BNO,L=London,C=GB"
+configPath = ./config.conf
+syncInterval = 9999999
+notary = "O=Notary,L=London,C=GB"
+bno = "O=BNO,L=London,C=GB"
